Move pages folder
