+++ frontend/src/pages/Admin/Admin.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import AuthContext from '../../context/auth-context';
+import AuthContext from '../../components/context/auth-context';
 import MentorItem from './MentorItem';
-import { fetchMentors } from '../../../api-calls/Mentors';
-import { fetchNoAuth, fetchAuth, fetchRun } from '../../../api-calls/Fetch';
-import Notification from '../../Notification/Notification';
-import Backdrop from '../../Backdrop/Backdrop';
-import Modal from '../../Modal/Modal';
+import { fetchMentors } from '../../api-calls/Mentors';
+import { fetchNoAuth, fetchAuth, fetchRun } from '../../api-calls/Fetch';
+import Notification from '../../components/Notification/Notification';
+import Backdrop from '../../components/Backdrop/Backdrop';
+import Modal from '../../components/Modal/Modal';
 
 import './Admin.css';
