+++ frontend/src/pages/Tasks/TaskItem.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { TASKS } from '../../../constants/tasks';
-import AuthContext from '../../context/auth-context';
+import { TASKS } from '../../constants/tasks';
+import AuthContext from '../../components/context/auth-context';
 import { NavLink } from 'react-router-dom';
-import { Free, Taken, NotStarted, InProgress, Done } from '../../Tags/Tags';
-import { fetchAuth } from '../../../api-calls/Fetch';
-import Backdrop from '../../Backdrop/Backdrop';
-import Modal from '../../Modal/Modal';
+import { Free, Taken, NotStarted, InProgress, Done } from '../../components/Tags/Tags';
+import { fetchAuth } from '../../api-calls/Fetch';
+import Backdrop from '../../components/Backdrop/Backdrop';
+import Modal from '../../components/Modal/Modal';
 
 import './TaskItem.css';
